Started working on a post jam port to desktop
There was like 3 hours (or so left... I'm exaggerating) so I started a windows port. I probably broke someting. Speaking of which, when it comes to browser. Never use ALT for anything in the future. Fortunately this game uses CTRL and ALT for the same thing. I didn't test on OSX. :(